renamed Button component file and removed old file
