+++ src/components/Buttons/Button.js
@@ -1,78 +1,69 @@
-// Library imports
 import React from "react";
 import { useNavigate } from "react-router-dom";
 
-/**
- * Component for providing a Button, provides a series of flexible subcomponents
- * that can be used to improve consistency and re-usability.
- * @param props
- * @returns {JSX.Element}
- * @constructor
- */
-const Button = (props) => {
-  // React hooks to redirect the user
+const Button = ({
+  className,
+  onClick,
+  children,
+  Icon,
+  path,
+  iconSize,
+  src,
+  srcId,
+  alt,
+  fill,
+  iconOnly,
+  disabled,
+  toggle,
+  active,
+  danger,
+  id,
+  onChange,
+}) => {
+  let buttonClassName = "button-basic " + (className ?? "");
+  buttonClassName += fill ? " button-fill" : "";
+  buttonClassName += iconOnly ? " button-icon" : "";
+  buttonClassName += disabled ? " button-disabled" : "";
+  buttonClassName += active ? " button-active" : "";
+  buttonClassName += danger ? " button-danger" : "";
+
   const navigate = useNavigate();
-  // Base CSS class to be used for buttons
-  let className = "standard-button ";
 
-  // Additional CSS class modifiers
-  className += props.className;
-  className += props.light ? " light-button " : ""
-  className += props.active ? " button__active" : "";
-  className += props.horizontal ? " horizontal-button" : " vertical-button";
-  className += props.checked ? " button__checked" : "";
-  className += props.border ? " button__border" : "";
-  className += props.accent ? " button__accent" : "";
-
-  // Get a list of DOM keys used for the Button component
-  let subComponentList = Object.keys(Button);
-  // Iterate through the list of child components provided
-  let subComponents = subComponentList.map((key) => {
-    return React.Children.map(props.children, (child) =>
-      child.type.displayName === key ? child : null
-    );
-  });
-
-  // Redirect the user to the provided path
-  const handlePath = () => {
-    navigate(props.path);
+  const handleClick = (e) => {
+    if (path && onClick) {
+      onClick(e);
+      navigate(path);
+    } else if (path && !onClick) {
+      navigate(path);
+    } else if (!path && onClick) {
+      onClick(e);
+    } else if (onChange) {
+      onChange(id);
+    }
   };
 
-  // Handle onClick path redirection
-  const onClick = props.path ? handlePath : props.onClick;
+  if (toggle) {
+    let imageClassName = " standard-image " + (active ?? "image-active");
+    let iconClassName = " toggle-icon " + (active ?? " icon-active ");
+
+    return (
+      <div className="button-toggle">
+        <button className={buttonClassName} onClick={handleClick} disabled={disabled}>
+          {src || srcId ? <img className={imageClassName} src={src} id={srcId} alt={alt} /> : <></>}
+          {Icon ? <Icon className={iconClassName} size={iconSize} /> : <></>}
+        </button>
+        {children ? <p>{children}</p> : <></>}
+      </div>
+    );
+  }
 
   return (
-    <button className={className} onClick={onClick} disabled={props.disabled}>
-      {subComponents.map((component) => component)}
+    <button id={id} className={buttonClassName} onClick={handleClick} disabled={disabled}>
+      {src || srcId ? <img className="standard-image" src={src} id={srcId} alt={alt} /> : <></>}
+      {Icon ? <Icon size={iconSize} /> : <></>}
+      {children ? <p>{children}</p> : <></>}
     </button>
   );
 };
-
-// Icon component for displaying icons in the button
-const Icon = (props) => {
-  let className = "button__icon " + props.className;
-  return <div className={className}>{props.children}</div>;
-};
-Icon.displayName = "Icon";
-Button.Icon = Icon;
-
-// Image component for displaying images in the button
-const Image = (props) => {
-  let className = "button__image " + props.className;
-  // Set the color scheme to active when the button is pressed or active
-  className += props.active ? " button__active" : "";
-
-  return <div className={className}>{props.children}</div>;
-};
-Image.displayName = "Image";
-Button.Image = Image;
-
-// Text component to allow the usage of text and text modification in the button
-const Text = (props) => {
-  let className = "button__text mb-0 " + props.className;
-  return <p className={className}>{props.children}</p>;
-};
-Text.displayName = "Text";
-Button.Text = Text;
 
 export default Button;
