+++ src/pages/Private/Workouts/sections/SectionAddWorkout.js
@@ -5,7 +5,7 @@ import { useNavigate } from "react-router-dom";
 
 // Local component imports
 import Section from "../../../../components/Misc/Section";
-import Button from "../../../../components/Buttons/buttons";
+import Button from "../../../../components/Buttons/Button";
 
 // Local services imports
 import services from "../../../../services/services";
